Move ECursor to GuiTypes.h, exclude TVirtualX.f from TVirtualPad.h
